LPD-90495 simplifies JiraAsset service structure, and adds entity deletion to JiraAssetSchemaService
